Java 25

+++ src/unsafe-accessor-unsafe-api/src/main/java/moe/karla/usf/unsafe/sunlegacy/Unsafe_Jdk8.java
@@ -477,7 +477,7 @@ class Unsafe_Jdk8 extends moe.karla.usf.unsafe.Unsafe {
     }
 
     @Override
-    public int arrayBaseOffset(Class<?> arrayClass) {
+    public long arrayBaseOffset(Class<?> arrayClass) {
         return theUnsafe.arrayBaseOffset(arrayClass);
     }
